Finessed the JS and styles to BANG
Looks good now. The elements get moved from the Header to the Nav, and vice versa if the viewport increased. A single breakpoint variable drives it all, with JS listening to a CSS custom prop value as its own breakpoint. Pretty cool if I do say so myself (and I just did)