Color of links inheriting from light mode
The color of links in light mode was too dark to be over a dark bg. We did not expect to deal with links here but we have them, so, the colors have been manually added to the theme. Also removed th eposition absolute for smaller than 640 screens, as long captions will completely cover the image for mobile